chore: change loadable function path
