+++ src/pages/careers/loading.ts
@@ -1,0 +1,7 @@
+import Loadable from '@loadable/component'
+
+export const WorkWithUs = Loadable(() => import('./work-with-us'))
+export const WhyWithUs = Loadable(() => import('./why-with-us'))
+export const PerksBenefits = Loadable(() => import('./perks-benefits'))
+export const OpenPositions = Loadable(() => import('./open-positions'))
+export const HiringProcess = Loadable(() => import('./hiring-process'))
