+++ src/pages/open-positions/loading.ts
@@ -1,0 +1,4 @@
+import Loadable from '@loadable/component'
+
+export const Hero = Loadable(() => import('./hero'))
+export const OpenPositions = Loadable(() => import('./open-positions'))
